+++ lmdeploy/pytorch/nn/linear/w8a8.py
@@ -42,7 +42,7 @@ class W8A8Linear(LinearBase):
                 in_features=in_features,
                 out_features=out_features,
                 bias=bias,
-                dtype=self.dtype,
+                output_dtype=self.dtype,
                 quant_dtype=quant_dtype,
             ),
             enable_deterministic=get_build_model_context().enable_deterministic,
