+++ lmdeploy/pytorch/nn/moe/static_fp8.py
@@ -258,7 +258,7 @@ class FusedMoEStaticF8(FusedMoEBase):
                 top_k=top_k,
                 num_experts=num_experts,
                 renormalize=renormalize,
-                out_dtype=dtype,
+                output_dtype=dtype,
                 quant_dtype=quant_dtype,
             ),
             enable_deterministic=get_build_model_context().enable_deterministic,
